+++ app/routes/decks/$id/export.tsx
@@ -1,10 +1,1 @@
-import { json } from "@remix-run/server-runtime";
-import { exportDeckJson } from "../../../misc/seed-utils";
-import { requireUserAndDeck } from "../../../utils/loader-deck-utils";
-import { makeLoader } from "../../../utils/loader-utils.server";
-
-export const loader = makeLoader(async ({ ctx }) => {
-  const { deck } = await requireUserAndDeck(ctx);
-  const data = await exportDeckJson(deck.id);
-  return json(data);
-});
+export { loader } from "./export.server";
