+++ app/routes/decks/new.server.tsx
@@ -1,0 +1,6 @@
+import { makeLoader } from "../../utils/loader-utils.server";
+
+export const loader = makeLoader(async ({ ctx }) => {
+  await ctx.requireUser();
+  return null;
+});
